requirements-dev.txt
pytest==6.2.5
flake8
black
boto3-stubs